backend/pyproject.toml
[project]
name = "analogstepik"
version = "0.1.0"
description = ""
authors = [
    {name = "grishka",email = "kolyackaam@gmail.com"}
]

readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask (>=3.1.3,<4.0.0)",
    "fastapi (>=0.136.0,<0.137.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "python-multipart (>=0.0.26,<0.0.27)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
